pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pyapicaller"
version = "0.2.1"
description = "Easily connect any API to OpenAI function calling"
authors = [{ name="Roman Medvedev", email="pypi@romavm.dev" }]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

keywords = ["API", "caller", "Python"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    'jsonref',
    ]

[project.urls]
homepage = "https://github.com/Romamo/pyapicaller"

[project.optional-dependencies]
yaml = ["pyyaml"]
swagger = ["urllib3 >= 1.15", "six >= 1.10", "certifi", "python-dateutil"]
openai = ["openai"]
